Cleaner works, adds updates to file

+++ DuplicateCleaner/Cleaner.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ProjectUtility;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -10,8 +11,6 @@ namespace DuplicateCleaner
 {
     internal class Cleaner
     {
-        private const float NAME_MATCH_THRESHOLD = 0.5f;//at least 50% match to count
-
         /// <summary>Gets the new company name from the project file</summary>
         /// <param name="path">path to the project file</param>
         /// <returns>company name if found, unknown if not</returns>
@@ -36,22 +35,35 @@ namespace DuplicateCleaner
             }
             return "Onbekend bedrijf";
         }
-
-        public string[] GetFilesInCompanyDir(string companyName)
+        public string[] GetFilesInCompanyDir(string companyName, out bool isNew)
         {
+            isNew = false;
             if (Util.IsOutputSet() == false)
             {
                 return new string[0];
             }
+            string dir = GetCompanyDirectory(companyName, out isNew);
+            if (isNew)
+            {//no directory found
+                Directory.CreateDirectory(dir);
+            }
+            return Directory.GetFiles(dir);
+        }
+        public string GetCompanyDirectory(string companyName, out bool isNew)
+        {
+            isNew = false;
+            if (Util.IsOutputSet() == false)
+            {
+                return string.Empty;
+            }
             string[] companies = Directory.GetDirectories(Settings.Instance.OutputPath);
             string[] companyNameParts = companyName.Split(' ');
-            string currentCompany = string.Empty;
-            int longestMatch = 0, matchIndex = 0;
+            int longestMatch = 0, matchIndex = -1;
             for (int i = 0; i < companies.Length; i++)
             {
                 if (Path.GetFileName(companies[i]).Equals(companyName, StringComparison.OrdinalIgnoreCase))
                 {
-                    return Directory.GetFiles(companies[i]);
+                    return companies[i];
                 }
                 string[] nameParts = Path.GetFileName(companies[i]).Split(' ');
                 int shortest = nameParts.Length < companyNameParts.Length ? nameParts.Length : companyNameParts.Length;
@@ -63,13 +75,17 @@ namespace DuplicateCleaner
                         {
                             longestMatch = p + 1;
                             matchIndex = i;
-                            currentCompany = Path.GetFileName(companies[i]);
                         }
                         break;
                     }
                 }
             }
-            return Directory.GetFiles(companies[matchIndex]);
+            if (matchIndex < 0)
+            {
+                isNew = true;
+                return Path.Combine(Settings.Instance.OutputPath, companyName);
+            }
+            return companies[matchIndex];
         }
 
         /// <summary>Tries to find matches to the file at path in the given file array</summary>
@@ -80,36 +96,22 @@ namespace DuplicateCleaner
         {
             if (files.Length == 0)
             { return files; }
-
             string projectName, compareName;
-            string[] projectNameParts, compareNameParts;
+            float maxLevenShtein = (float)Settings.Instance.SearchAccuracy * 0.01f; ;
             List<string> possibleMatches = new List<string>();
 
             //TODO: put regex here to remove date from title ("Aanvraag WBSO xxxx")
             projectName = TrimExtractionData(File.ReadAllLines(path)[0]);
-            projectNameParts = projectName.Split(' ');
+            projectName = TrimToLastInvalid(projectName);
 
             for (int i = 0; i < files.Length; i++)
             {
                 compareName = TrimExtractionData(File.ReadAllLines(files[i])[0]);
-                if (projectName.Equals(compareName, StringComparison.OrdinalIgnoreCase))
-                {//exact match found is duplicate, return only that
-                    return new string[] { files[i] };
-                }
+                compareName = TrimToLastInvalid(compareName);
                 //TODO: put regex here to remove date from title ("Aanvraag WBSO xxxx")
-                compareNameParts = compareName.Split(' ');
-                int shortest = compareNameParts.Length < projectNameParts.Length ? compareNameParts.Length : projectNameParts.Length;
-                float match = 0;
-                for (int p = 0; p < shortest; p++)
-                {
-                    int compareP = compareNameParts.Length - 1 - p;
-                    int projectP = projectNameParts.Length - 1 - p;
-                    if (compareNameParts[compareP].Equals(projectNameParts[projectP], StringComparison.OrdinalIgnoreCase))
-                    {
-                        match = ((float)p + 1) / (float)shortest;
-                    }
-                }
-                if (match >= NAME_MATCH_THRESHOLD)
+                int dist = StringUtil.Levenshtein(projectName, compareName);
+                float match = (float)dist / (float)compareName.Length;
+                if (match <= maxLevenShtein)
                 {
                     possibleMatches.Add(files[i]);
                 }
@@ -117,6 +119,26 @@ namespace DuplicateCleaner
             return possibleMatches.ToArray();
         }
 
+        public void UpdateWithMatches(string[] files, string path)
+        {
+            bool hasUpdates = FileContainsUpdates(path, out int index);
+            string[] updates = new string[files.Length];
+            for (int i = 0; i < files.Length; i++)
+            {
+                updates[i] = TryFindUpdates(files[i]);
+            }
+            if (!hasUpdates)
+            {
+                AfterDescription(path, out index);
+            }
+            string final = InsertUpdate(path, string.Join(Environment.NewLine, updates), index, !hasUpdates);
+            using (StreamWriter sw = File.CreateText(path))
+            {
+                //write the final result to a text document
+                sw.Write(final);
+                sw.Close();
+            }
+        }
         public void ReplaceProject(string pathToOld, string pathToNew, bool moveNew = true)
         {
             string dir, oldProjDir, oldDestination, newDestionation;
@@ -131,10 +153,89 @@ namespace DuplicateCleaner
             File.Move(pathToOld, oldDestination);
             if (moveNew)
             {
+                if (File.Exists(newDestionation))
+                {
+                    Console.WriteLine("File exists at destination, skipping... ({0})", newDestionation);
+                    return;
+                }
                 File.Move(pathToNew, newDestionation);
             }
         }
 
+        private string TryFindUpdates(string path)
+        {
+            if (!FileContainsUpdates(path, out int index))
+            { return string.Empty; }
+
+            string[] lines = File.ReadAllLines(path);
+            StringBuilder str = new StringBuilder();
+            for (int i = index; i < lines.Length; i++)
+            {
+                if (lines[i].StartsWith("update project:", StringComparison.OrdinalIgnoreCase))
+                {
+                    continue;
+                }
+                if (string.IsNullOrWhiteSpace(lines[i]))
+                {
+                    break;
+                }
+                str.AppendLine(lines[i]);
+            }
+            return "-" + str.ToString().Trim();
+        }
+        private bool FileContainsUpdates(string path, out int index)
+        {
+            index = -1;
+            string[] lines = File.ReadAllLines(path);
+            for (int i = 0; i < lines.Length; i++)
+            {
+                if (lines[i].StartsWith("update project:", StringComparison.OrdinalIgnoreCase))
+                {
+                    index = i;
+                    return true;
+                }
+            }
+            return false;
+        }
+        private bool AfterDescription(string path, out int index)
+        {
+            index = -1;
+            string[] lines = File.ReadAllLines(path);
+            bool description = false;
+            for (int i = 0; i < lines.Length; i++)
+            {
+                if (lines[i].StartsWith("update project:", StringComparison.OrdinalIgnoreCase))
+                {
+                    description = true;
+                    continue;
+                }
+                if (description == true && string.IsNullOrWhiteSpace(lines[i]))
+                {
+                    index = i;
+                    break;
+                }
+            }
+            return description;
+        }
+        private string InsertUpdate(string path, string toAdd, int index, bool addUpdateText)
+        {
+            List<string> file = new List<string>();
+            file.AddRange(File.ReadAllLines(path));
+            if (index + 1 > file.Count)
+            {
+                file.Add(toAdd);
+            }
+            else
+            {
+                string final = toAdd;
+                if (addUpdateText)
+                {
+                    final = "Update Project:" + Environment.NewLine + final;
+                }
+                file.Insert(index + 1, final);
+            }
+            return string.Join(Environment.NewLine, file);
+        }
 
         private string TrimExtractionData(string name)
         {
@@ -144,6 +245,18 @@ namespace DuplicateCleaner
             }
             name = name.Trim();
             return name;
+        }
+        private string TrimToLastInvalid(string value)
+        {
+            List<char> ill = Path.GetInvalidFileNameChars().ToList();
+            ill.Add('-');
+            ill.Add('_');
+            int last = value.LastIndexOfAny(ill.ToArray());
+            if (last < 0)
+            {
+                return value;
+            }
+            return value.Substring(last);
         }
     }
 }
